Removed InputHelper and added a StringExtension instead, also rearanged file structure in the project
