--- Assets/Events/Events.cs
@@ -1,7 +1,0 @@
-using System;
-using UnityEngine.Events;
-
-namespace FG {
-	[Serializable]
-		public class FloatEvent : UnityEvent<float> { }
-}
