--- Assets/Events/Events.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 942213bd98b5cd146b0a2120abf321b9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
