+++ Assets/Prefabs/Cars/Big Boi.prefab
@@ -336,7 +336,7 @@ GameObject:
   - component: {fileID: 9081775713173241854}
   - component: {fileID: 6479560205463259956}
   m_Layer: 0
-  m_Name: CarBus
+  m_Name: Big Boi
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
