+++ Assets/Prefabs/Player.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0659751dfdc7c544fbca2ee93f4b2105
+guid: f96ad7aff8dbb5b409cdcafd2e0d3581
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
