+++ Assets/Prefabs/Player/Player.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 585c8a7e7cb08da49bffba30583f3d41
-NativeFormatImporter:
+guid: b7825c452dabdd546b2f37d4c23797a1
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
