+++ Assets/Scripts/Extensions.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 897e3b80264dd704cac01f0c11a6a3a3
-PrefabImporter:
+guid: d0f5c676d847ae54b988151c9ee4d2b4
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
