+++ Assets/Scripts/Extensions/StringExtensions.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 8fdd85ea260b4fd2b187bbe881e76a11
+timeCreated: 1629989638
