--- Assets/Scripts/InputHelper.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 8fdd85ea260b4fd2b187bbe881e76a11
-timeCreated: 1629989638
