+++ Assets/Scripts/Player.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: c713dd67096db594e80e8fe411b4a22e
-PrefabImporter:
+guid: 99cbe31ac8118f14b993de36a8b77cf2
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
